+++ Posit/Posit32_2.cs
@@ -854,14 +854,14 @@ namespace Lombiq.Arithmetics
 			if (resultFractionBits == 0) return new Posit32_2(0, true);
 
 			var resultRegimeKValue = scaleFactor / (1 << MaximumExponentSize);
-			var resultExponentBits = (uint)(scaleFactor % (1 << MaximumExponentSize));
+			var resultExponentBits = (scaleFactor % (1 << MaximumExponentSize));
 			if (resultExponentBits < 0)
 			{
 				resultRegimeKValue -= 1;
-				resultExponentBits += (uint)(1 << MaximumExponentSize);
+				resultExponentBits += (1 << MaximumExponentSize);
 			}
 
-			return new Posit32_2(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue, resultExponentBits, resultFractionBits), true);
+			return new Posit32_2(AssemblePositBitsWithRounding(resultSignBit, resultRegimeKValue,(uint) resultExponentBits, resultFractionBits), true);
 		}
 
 		public static Posit32_2 operator +(Posit32_2 left, int right) => left + new Posit32_2(right);
